Move `BitWarden` class to `passmngt.py` module
A new `passmngt.py` can be used to handle all password managers.
