+++ gcexport.py
@@ -23,7 +23,7 @@ from datetime import datetime, timedelta, tzinfo
 from getpass import getpass
 from math import floor
 from platform import python_version
-from subprocess import call, run
+from subprocess import call
 from timeit import default_timer as timer
 
 import argparse
@@ -40,6 +40,7 @@ import unicodedata
 import zipfile
 
 from filtering import update_download_stats, read_exclude
+from passmngt import BitWarden
 
 import http.cookiejar
 import urllib.error
@@ -1141,56 +1142,6 @@ def copy_details_to_summary(summary, details):
     summary['maxHR'] = details['summaryDTO']['maxHR'] if 'summaryDTO' in details and 'maxHR' in details['summaryDTO'] else None
     summary['averageHR'] = details['summaryDTO']['averageHR'] if 'summaryDTO' in details and 'averageHR' in details['summaryDTO'] else None
     summary['elevationCorrected'] = details['metadataDTO']['elevationCorrected'] if 'metadataDTO' in details and 'elevationCorrected' in details['metadataDTO'] else None
-
-
-class BitWarden:
-    META = "BW_SESSION="
-
-    def __init__(self, url="garmin.com"):
-        self.user = None
-        self.password = None
-        self.session = None
-        self.url = url
-
-    def lock(self):
-        self.user = None
-        self.password = None
-        self.session = None
-        run(["bw", "lock"])
-
-    def __enter__(self):
-        print("Enter your master password: ", end="", flush=True)
-        bw = run(["bw", "unlock"], capture_output=True)
-        print()
-
-        output = bw.stdout.decode()
-        try:
-            session_start = output.index(self.META) + len(self.META)
-            session_stop = output.index('"', session_start + 1)
-        except ValueError:
-            logging.error(bw.stderr.decode())
-            self.lock()
-            sys.exit(1)
-
-        self.session = output[session_start:session_stop]
-        self.user = run(
-            ["bw", "get", "username", self.url, "--session", self.session],
-            capture_output=True,
-        ).stdout.decode()
-        self.password = run(
-            ["bw", "get", "password", self.url, "--session", self.session],
-            capture_output=True,
-        ).stdout.decode()
-
-        if not all((self.user, self.password)):
-            logging.error("BitWarden cannot find credentials for: %s", self.url)
-            self.lock()
-            sys.exit(1)
-
-        return self
-
-    def __exit__(self, exc_type, exc_value, exc_traceback):
-        self.lock()
 
 
 def main(argv):
